Coordinated-Multi-Agent-System/pyproject.toml
[project]
name = "openai-agent-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
